# Rewrite the compiler.