@BurntSushi 2016-01-10
Before:

```
test bench::anchored_literal_long_match      ... bench:         156 ns/iter (+/- 5)
test bench::anchored_literal_long_non_match  ... bench:          85 ns/iter (+/- 7)
test bench::anchored_literal_short_match     ... bench:         145 ns/iter (+/- 3)
test bench::anchored_literal_short_non_match ... bench:          76 ns/iter (+/- 2)
test bench::easy0_1K                         ... bench:         269 ns/iter (+/- 63) = 3806 MB/s
test bench::easy0_1MB                        ... bench:     232,461 ns/iter (+/- 13,022) = 4509 MB/s
test bench::easy0_32                         ... bench:          63 ns/iter (+/- 6) = 507 MB/s
test bench::easy0_32K                        ... bench:       8,358 ns/iter (+/- 430) = 3920 MB/s
test bench::easy1_1K                         ... bench:         274 ns/iter (+/- 101) = 3737 MB/s
test bench::easy1_1MB                        ... bench:     278,949 ns/iter (+/- 11,324) = 3758 MB/s
test bench::easy1_32                         ... bench:          63 ns/iter (+/- 15) = 507 MB/s
test bench::easy1_32K                        ... bench:       7,731 ns/iter (+/- 488) = 4238 MB/s
test bench::hard_1K                          ... bench:      44,685 ns/iter (+/- 661) = 22 MB/s
test bench::hard_1MB                         ... bench:  60,108,237 ns/iter (+/- 814,810) = 16 MB/s
test bench::hard_32                          ... bench:       1,412 ns/iter (+/- 38) = 22 MB/s
test bench::hard_32K                         ... bench:   1,363,335 ns/iter (+/- 21,316) = 24 MB/s
test bench::literal                          ... bench:          14 ns/iter (+/- 0)
test bench::match_class                      ... bench:          81 ns/iter (+/- 0)
test bench::match_class_in_range             ... bench:          81 ns/iter (+/- 2)
test bench::match_class_unicode              ... bench:       2,978 ns/iter (+/- 64)
test bench::medium_1K                        ... bench:       2,239 ns/iter (+/- 68) = 457 MB/s
test bench::medium_1MB                       ... bench:   2,215,729 ns/iter (+/- 20,897) = 472 MB/s
test bench::medium_32                        ... bench:         124 ns/iter (+/- 2) = 258 MB/s
test bench::medium_32K                       ... bench:      72,486 ns/iter (+/- 1,027) = 452 MB/s
test bench::no_exponential                   ... bench:     282,992 ns/iter (+/- 8,102)
test bench::not_literal                      ... bench:       1,526 ns/iter (+/- 32)
test bench::one_pass_long_prefix             ... bench:         307 ns/iter (+/- 7)
test bench::one_pass_long_prefix_not         ... bench:         311 ns/iter (+/- 8)
test bench::one_pass_short_a                 ... bench:         623 ns/iter (+/- 12)
test bench::one_pass_short_a_not             ... bench:         920 ns/iter (+/- 19)
test bench::one_pass_short_b                 ... bench:         554 ns/iter (+/- 13)
test bench::one_pass_short_b_not             ... bench:         740 ns/iter (+/- 12)
test bench::replace_all                      ... bench:         155 ns/iter (+/- 5)
```

After:

```
test bench::anchored_literal_long_match      ... bench:         145 ns/iter (+/- 1)
test bench::anchored_literal_long_non_match  ... bench:          92 ns/iter (+/- 2)
test bench::anchored_literal_short_match     ... bench:         129 ns/iter (+/- 3)
test bench::anchored_literal_short_non_match ... bench:          72 ns/iter (+/- 1)
test bench::easy0_1K                         ... bench:         268 ns/iter (+/- 88) = 3820 MB/s
test bench::easy0_1MB                        ... bench:     234,067 ns/iter (+/- 4,663) = 4479 MB/s
test bench::easy0_32                         ... bench:          64 ns/iter (+/- 4) = 500 MB/s
test bench::easy0_32K                        ... bench:       8,298 ns/iter (+/- 521) = 3948 MB/s
test bench::easy1_1K                         ... bench:         275 ns/iter (+/- 95) = 3723 MB/s
test bench::easy1_1MB                        ... bench:     280,466 ns/iter (+/- 5,938) = 3738 MB/s
test bench::easy1_32                         ... bench:          64 ns/iter (+/- 16) = 500 MB/s
test bench::easy1_32K                        ... bench:       7,693 ns/iter (+/- 595) = 4259 MB/s
test bench::hard_1K                          ... bench:      27,844 ns/iter (+/- 1,012) = 36 MB/s
test bench::hard_1MB                         ... bench:  52,323,489 ns/iter (+/- 1,251,665) = 19 MB/s
test bench::hard_32                          ... bench:         970 ns/iter (+/- 92) = 32 MB/s
test bench::hard_32K                         ... bench:     896,945 ns/iter (+/- 29,977) = 36 MB/s
test bench::literal                          ... bench:          13 ns/iter (+/- 1)
test bench::match_class                      ... bench:          80 ns/iter (+/- 0)
test bench::match_class_in_range             ... bench:          80 ns/iter (+/- 0)
test bench::match_class_unicode              ... bench:       2,150 ns/iter (+/- 18)
test bench::medium_1K                        ... bench:       2,241 ns/iter (+/- 55) = 456 MB/s
test bench::medium_1MB                       ... bench:   2,186,354 ns/iter (+/- 9,134) = 479 MB/s
test bench::medium_32                        ... bench:         125 ns/iter (+/- 1) = 256 MB/s
test bench::medium_32K                       ... bench:      72,156 ns/iter (+/- 145) = 454 MB/s
test bench::no_exponential                   ... bench:     305,034 ns/iter (+/- 1,134)
test bench::not_literal                      ... bench:       1,169 ns/iter (+/- 105)
test bench::one_pass_long_prefix             ... bench:         257 ns/iter (+/- 4)
test bench::one_pass_long_prefix_not         ... bench:         276 ns/iter (+/- 4)
test bench::one_pass_short_a                 ... bench:         680 ns/iter (+/- 3)
test bench::one_pass_short_a_not             ... bench:         804 ns/iter (+/- 48)
test bench::one_pass_short_b                 ... bench:         337 ns/iter (+/- 3)
test bench::one_pass_short_b_not             ... bench:         339 ns/iter (+/- 5)
test bench::replace_all                      ... bench:         150 ns/iter (+/- 1)
```

From what I can see, there are no regressions. There are some nice perf wins, and I didn't expect them to be this big. The `hard` benchmark (which defeats all optimizations) and other regexes, like the `one_pass` regexes, seem to benefit the most. The `one_pass_short_b` regex in particular improves nicely. The regex is `^.bc(?:d|e)*$` Here is what the regex compiled to before:

```
000 Save(0)
001 StartText
002 '\u{0}'-'\t', '\u{b}'-'\u{10ffff}'
003 'b'
004 'c'
005 Split(6, 11)
006 Split(7, 9)
007 'd'
008 Jump(10)
009 'e'
010 Jump(5)
011 EndText
012 Save(1)
013 Match
```

And after:

```
000 Save(0)
001 StartText
002 '\u{0}'-'\t', '\u{b}'-'\u{10ffff}'
003 'b'
004 'c'
005 Split(6, 9)
006 Split(7, 8)
007 'd' (goto: 5)
008 'e' (goto: 5)
009 EndText
010 Save(1)
011 Match
```

Both of the `Jump` instructions are eliminated, which are probably otherwise executed a lot. Removing this indirection means fewer function calls.