@BurntSushi 2016-01-10
I'm not sure who to ask for review other than @alexcrichton, but I know @WillEngler is familiar with the code and @jneem is doing work in this space too.