@BurntSushi 2016-01-11
@jneem Thank you so much for doing a code review! I think I like all of your ideas. I might not get to them for a couple days though!