@alexcrichton 2016-01-11
At a cursory glance this looks good to me, although I think I'm not even qualified to give this a thorough review :). I'd be willing to merge as soon as you're ready @BurntSushi, or feel free to just merge yourself!