@BurntSushi 2016-01-16
Thanks for comments and review! I've obliged all of @jneem's suggestions. :-)